requirements.txt
# SOC-SLM Requirements
# Core dependencies for the Small Language Model

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# NLP & Tokenization
regex>=2023.0.0
sentencepiece>=0.1.99

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging & Monitoring
structlog>=23.0.0

# Optional: Distributed Training
# accelerate>=0.20.0
# deepspeed>=0.10.0

# Optional: Advanced Quantization
# bitsandbytes>=0.41.0
# auto-gptq>=0.4.0

# Optional: Flash Attention (requires CUDA)
# flash-attn>=2.0.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.15.0
